POC Dockerize Local Dev
After trying a handful of times to get this running locally on macOS
both bare and virtualbox, I decided that it would be just be easier to
dockerize this so people and clone, build and run this with 3 commands.

__NOTE:__ this is not master ready! This still needs thorough testing
and production deployment merges so as not to break anything. This is a
proof of concept for local dev only.

Changes:

- moved all app code to `/app`
- renamed a couple environment variables to suit docker conventions
- removed vagrant files and information from readme
- updated readme to instruct users on how to use this with docker

Caveats:

- i'm still super new to this app, so this hasn't been thoroughly
tested: read this as __it will probably break somewhere__
- depending on how this app is deployed to production, this will need a
ton of updates to not break deployment conventions