Reverted Marshmallow Version
Newest version of `marshmallow` library strictly enforces datetime
formatting of query params -- that is to say `"2015-1-1"` is not a valid
string representation of a datetime and thus bombs the tests.

This fixes that issue, but I don't like it. We should be strictly
enforcing proper ISO-8601 date time strings. It's a bug, not a feature.

Also added `nose` to the requirements.