It's Not a Bug, It's a CVE as a Feature!
So one of the things that I hate the most about this app's API is that
timestamps are expressed as `YYYY-M-D` instead of `YYYY-MM-DD`. This is
a bug that existed in marshmallow and was subsequently removed. However,
a massive swath of the tests use this format, as do the docs, example
code, and almost all the logged API requests follow that pattern (jesus
christ why!!!?).

Upgrading the marshmallow dependency to a safe version breaks all of
this. This commit rolls that back to the version that "works".